Fix interestingness usage
The `interestingness` parameter was being used only to initialize the
first `next_node` and nowhere else.  Also, the way it was used was
flawed since it was picking the first key of the dictionary, which is
not guaranteed to be the one with the highest interest.

This commit properly uses `interestingness` both for choosing the first
node and to pick the best candidate for `next_node`, when more than on
is available.